routes added and refactoring of views-no obvious bugs
