+++ src/components/Nav/Nav.module.scss
@@ -24,5 +24,4 @@
 .Favorites {
     text-decoration: none;
     color: white;
-    margin-left: 60%;
 }
